correlation surface: move internal methods to a separate file to avoid pulling pyzx and solve circular imports
